backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-pptx==1.0.2
anthropic==0.40.0
playwright==1.48.0
docling==2.15.0
Pillow==10.4.0
python-multipart==0.0.12
aiofiles==24.1.0
sqlalchemy==2.0.35
aiosqlite==0.20.0
pydantic==2.9.0
